+++ sklearn/neighbors/classification.py
@@ -29,7 +29,7 @@ class KNeighborsClassifier(NeighborsBase, KNeighborsMixin,
     Parameters
     ----------
     n_neighbors : int, optional (default = 5)
-        Number of neighbors to use by default for :meth:`k_neighbors` queries.
+        Number of neighbors to use by default for :meth:`kneighbors` queries.
 
     weights : str or callable, optional (default = 'uniform')
         weight function used in prediction.  Possible values: